Use uniid from session as cardnumber
Instead of solving with a temporary hardcoded cardnumber
