+++ .env.example
@@ -18,7 +18,6 @@ GO_SESSION_SECRET=XXX
 NEXT_PUBLIC_APP_URL=https://dapple-cms.docker:3000
 
 # PUBHUB
-TMP_PUBLIZON_GET_LIBRARY_USER_ORDER_LIST_CARD_NUMBER=goxx0020
 UNLILOGIN_PUBHUB_CLIENT_ID=XXX
 UNLILOGIN_PUBHUB_RETAILER_ID=XXX
 UNLILOGIN_PUBHUB_RETAILER_KEY_CODE=XXX
